Updated security on some requests

+++ src/main/java/com/training/cinemaapp/controllers/BookingController.java
@@ -4,8 +4,10 @@ import com.training.cinemaapp.models.Booking;
 import com.training.cinemaapp.services.BookingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+import com.training.cinemaapp.security.UserSecurity;
 
 import javax.validation.Valid;
 import java.net.URI;
@@ -16,6 +18,9 @@ public class BookingController {
 
     @Autowired
     private BookingService bookingService;
+
+    @Autowired
+    private UserSecurity userSecurity;
 
     @GetMapping("/bookings")
     public List<Booking> getBookings() {
@@ -28,6 +33,7 @@ public class BookingController {
     }
 
     @GetMapping("/bookings/user/{userId}")
+//    @PreAuthorize("hasRole('ROLE_ADMIN') or @userSecurity.isUserBookingAuthor(authentication.principal.id, #userId)")
     public List<Booking> getBookingsByUserId(@PathVariable long userId) {
         return bookingService.getBookingsByUserId(userId);
     }
@@ -43,16 +49,22 @@ public class BookingController {
 
     @PostMapping("/bookings")
     public ResponseEntity<Booking> addBooking(@Valid @RequestBody Booking booking) {
-        Booking newBooking = bookingService.addBooking(booking);
-        URI location = ServletUriComponentsBuilder
-                .fromCurrentRequest()
-                .path("/{id}")
-                .buildAndExpand(newBooking.getId())
-                .toUri();
-        return ResponseEntity.created(location).body(newBooking);
+        //check if user is admin or is booking for himself
+        if (userSecurity.isUserBookingForHimself(booking.getUserId())) {
+            Booking newBooking = bookingService.addBooking(booking);
+            URI location = ServletUriComponentsBuilder
+                    .fromCurrentRequest()
+                    .path("/{id}")
+                    .buildAndExpand(newBooking.getId())
+                    .toUri();
+            return ResponseEntity.created(location).body(newBooking);
+        } else {
+            return ResponseEntity.status(403).build();
+        }
     }
 
     @PutMapping("/bookings/{id}")
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     public ResponseEntity<?> updateBooking(@Valid @RequestBody Booking newBooking, @PathVariable int id) {
         if (bookingService.updateBooking(newBooking, id).isPresent()) {
             return ResponseEntity.ok().body(bookingService.updateBooking(newBooking, id));
@@ -62,11 +74,19 @@ public class BookingController {
     }
 
     @DeleteMapping("/bookings/{id}")
+//    @PreAuthorize("hasRole('ROLE_ADMIN') or @userSecurity.isUserBookingAuthor(authentication.principal.id, #id)")
     public ResponseEntity<?> deleteBooking(@PathVariable int id) {
-        if (bookingService.deleteBooking(id)) {
-            return ResponseEntity.noContent().build();
+        if (userSecurity.isUserBookingAuthor(id)) {
+            if (bookingService.deleteBooking(id)) {
+                return ResponseEntity.noContent().build();
+            } else {
+                return ResponseEntity.notFound().build();
+            }
         } else {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(403).build();
         }
     }
+
 }
+
+
